rename icon folder from 'icon' to 'icons'
