+++ src/components/JdsIcon/Icon.vue
@@ -97,7 +97,7 @@ export default {
         return
       }
       // dynamic import component based on provided `name` prop
-      const iconComponent = await import(`../../assets/icon/${name}.svg?vue`)
+      const iconComponent = await import(`../../assets/icons/${name}.svg?vue`)
         .then((m) => m.default || m)
         // return null on import error
         .catch(() => null)
